requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
WTForms==3.0.1
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.0.0
Pillow==10.0.1
reportlab==4.0.4
openpyxl==3.1.2
